Add `id` field
I think it would often be convenient to have some ID for a given PEtab problem and I would like to add such a field to the yaml schema.
Previously, we usually used the model ID in such situations. However, often we have multiple problems based on the same model but different datasets, and this doesn't work well in the case of multiple models.

Not sure whether it should be mandatory or optional. I think, I'd go for optional.