Add `id` field (#646)
It is often convenient to have some string to identify a given PEtab problem. Previously, we usually used the model ID in such situations. However, often we have multiple problems based on the same model but different datasets, and this doesn't work well in the case of multiple models. Therefore, we add an optional `id` field to the PEtab problem schema.